Consolidate TrainingData and DetectorDataset class into one file.
